+++ backend/src/neuroagent/app/app_utils.py
@@ -4,11 +4,15 @@ import logging
 from typing import Any
 
 from fastapi import HTTPException
+from openai import AsyncOpenAI
+from openai.types.embedding import Embedding
+from sklearn.metrics.pairwise import cosine_similarity
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from starlette.status import HTTP_401_UNAUTHORIZED
 
 from neuroagent.app.config import Settings
+from neuroagent.tools.base_tool import BaseTool
 
 logger = logging.getLogger(__name__)
 
@@ -72,3 +76,47 @@ def validate_project(
     else:
         # No vlab nor project provided, nothing to do.
         return
+
+
+async def get_embeddings(
+    openai_client: AsyncOpenAI,
+    to_embed: list[str] | str,
+    embedding_model: str,
+    embedding_dim: int,
+) -> list[Embedding]:
+    """Embed tool description for later retrieval."""
+    embeddings = await openai_client.embeddings.create(
+        input=to_embed, model=embedding_model, dimensions=embedding_dim
+    )
+    return embeddings.data
+
+
+async def retrieve_tools(
+    tools_embedding_dict: dict[str, list[float]],
+    embedded_query: list[float],
+    tool_list: list[type[BaseTool]],
+    max_tools: int,
+) -> list[type[BaseTool]]:
+    """Get the top-k most relevant tools given a user query."""
+    # Get the cosine similarity between tools and query
+    tool_query_similarity = cosine_similarity(
+        [embedded_query], list(tools_embedding_dict.values())
+    ).squeeze(axis=0)
+
+    # Sort the tools based on their similarity with the query
+    sorted_tool_name, sorted_tool_scores = zip(
+        *sorted(
+            [
+                (tool_name, score)
+                for tool_name, score in zip(
+                    tools_embedding_dict.keys(), tool_query_similarity
+                )
+            ],
+            key=lambda x: x[1],
+            reverse=True,
+        )
+    )
+    logger.info(
+        f"Selected tools: {', '.join(sorted_tool_name[:max_tools])} with score {sorted_tool_scores[:max_tools]}"
+    )
+    return [tool for tool in tool_list if tool.name in sorted_tool_name[:max_tools]]